src/requirements.txt
ecdsa==0.18.0
msgpack_python==0.5.6
pycryptodomex==3.18.0
x25519==0.0.2
zstandard==0.19.0
